Move postsolidfilter to filter/postsolidfilter
