+++ pkg/protocol/engine/filter/postsolidfilter/postsolidblockfilter/post_solid_block_filter.go
@@ -8,7 +8,7 @@ import (
 	"github.com/iotaledger/iota-core/pkg/protocol/engine"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/accounts"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/blocks"
-	"github.com/iotaledger/iota-core/pkg/protocol/engine/postsolidfilter"
+	"github.com/iotaledger/iota-core/pkg/protocol/engine/filter/postsolidfilter"
 	iotago "github.com/iotaledger/iota.go/v4"
 )
